+++ packages/electron/src/main/ipc.js
@@ -2,13 +2,23 @@
 const fs = require('fs');
 const { resolve } = require('path');
 const { app, ipcMain, dialog, BrowserWindow } = require('electron');
-const esmRequire = require('esm')(module);
-const { getDefaultConfig } = esmRequire('../../../frontend/utils/config');
+const { updateTrayIcon } = require('./tray');
 
-const ipc = (server, options) => {
-  ipcMain.handle('startServer', async (event, port) => {
+const registerWindow = window => {
+  window.addListener('enter-full-screen', () => {
+    window.webContents.send('window', 'fullscreen', true);
+  });
+  window.addListener('leave-full-screen', () => {
+    window.webContents.send('window', 'fullscreen', false);
+  });
+};
+
+const ipc = (server) => {
+  ipcMain.handle('startServer', async (event, port, ssl) => {
+    console.log(port, ssl);
     try {
-      return await server.start(port);
+      await server.start(port, ssl);
+      return { ...server.toJSON() };
     } catch (error) {
       server.stop();
       return error;
@@ -24,6 +34,9 @@ const ipc = (server, options) => {
   ipcMain.handle('getServerOptions', (event) => {
     return getServerOptions(server);
   });
+  ipcMain.handle('serverSupported', (event) => {
+    return server.supported();
+  });
 
   const userDataPath = app.getPath('userData');
   const configFile = resolve(userDataPath, 'config.json');
@@ -33,7 +46,7 @@ const ipc = (server, options) => {
     try {
       return JSON.parse(await fs.promises.readFile(configFile, 'utf-8'));
     } catch (error) {
-      return getDefaultConfig();
+      return null;
     }
   };
   ipcMain.handle('loadConfig', onLoadConfig);
@@ -43,6 +56,18 @@ const ipc = (server, options) => {
     fs.promises.writeFile(configFile, JSON.stringify(data), 'utf-8');
   };
   ipcMain.handle('saveConfig', onSaveConfig);
+
+  const onClient = (event, type, value) => {
+    switch (type) {
+      case 'connect':
+        updateTrayIcon({ remote: value });
+        break;
+
+      default:
+        break;
+    }
+  };
+  ipcMain.handle('client', onClient);
 
   // Window
   const onWindow = (event, type, value) => {
@@ -114,17 +139,6 @@ const ipc = (server, options) => {
     return data;
   };
   ipcMain.handle('load', onLoad);
-
-  return {
-    registerWindowEvents: (window) => {
-      window.addListener('enter-full-screen', () => {
-        window.webContents.send('window', 'fullscreen', true);
-      });
-      window.addListener('leave-full-screen', () => {
-        window.webContents.send('window', 'fullscreen', false);
-      });
-    }
-  };
 };
 
 const getServerOptions = (server) => {
@@ -132,8 +146,10 @@ const getServerOptions = (server) => {
     active: server.active,
     port: server.port,
     hosts: server.hosts,
+    ssl: server.ssl,
+    secure: server.secure,
     activeSessions: server.sockets.size
   };
 };
 
-module.exports = { default: ipc };
+module.exports = { registerWindow, default: ipc };