+++ packages/frontend/components/views/Start.vue
@@ -2,7 +2,7 @@
   <app-view class="view-start">
     <div>
       <h2>Welcome DevTerm User!</h2>
-      <p v-if="'$server' in this">
+      <p v-if="!$server.disabled">
         Do you work locally or remotely?
       </p>
       <p v-else>
@@ -11,12 +11,8 @@
       <p>
         Click on "Use Remote" and enter the connection data <br>to connect to the DevTerm.
       </p>
-      <div>
-        <input-check-box v-model="remember" label="Remember choice" delimiter="?" />
-      </div>
-
       <div class="buttons">
-        <input-text-button v-if="'$server' in this" color="primary" text="Use Local" @click="onClickLocal" />
+        <input-text-button v-if="!$server.disabled" color="primary" text="Use Local" @click="onClickLocal" />
         <input-text-button color="primary" text="Use Remote" @click="onClickRemote" />
       </div>
     </div>
@@ -24,28 +20,23 @@
 </template>
 
 <script>
+
 import AppView from '../app/View.vue';
 import InputTextButton from '../inputs/TextButton.vue';
-import InputCheckBox from '../inputs/CheckBox.vue';
+
 export default {
   components: {
     AppView,
-    InputTextButton,
-    InputCheckBox
-  },
-  data () {
-    return { remember: false };
+    InputTextButton
   },
   methods: {
     onClickLocal () {
       this.$emit('apply', {
-        remember: this.remember,
         type: 'local'
       });
     },
     onClickRemote () {
       this.$emit('apply', {
-        remember: this.remember,
         type: 'remote'
       });
     }