+++ packages/frontend/plugins/server.web.js
@@ -1,0 +1,13 @@
+
+import Vue from 'vue';
+
+Object.defineProperty(Vue.prototype, '$server', {
+  get () {
+    return {
+      init () {
+        return Promise.resolve();
+      },
+      disabled: true
+    };
+  }
+});