+++ packages/web/src/main.js
@@ -5,8 +5,9 @@ import App from '@devterm-toolbox/frontend/components/App.vue';
 // plugins
 import '@devterm-toolbox/frontend/plugins/error';
 import '@devterm-toolbox/frontend/plugins/config';
-// import '@devterm-toolbox/frontend/plugins/server';
+import '@devterm-toolbox/frontend/plugins/server.web';
 import '@devterm-toolbox/frontend/plugins/client';
+import '@devterm-toolbox/frontend/plugins/dialog';
 
 Vue.component('App', App);